Organized more methods in `GtkExtensions`
